Limit write_to_robot payload to 64 bytes. Debug msgs end in \n
printf's used for debugging don't show up until a \n is sent. E.g. printf("hello"); shows nothing at all. printf(" world\n"); then shows "hello world". Changed the debug msgs to always terminate with a \n 

Because of overhead sent before the actual command oplet, the actual usable "max_content_chars" (in DDE) or "WRITE_TO_ROBOT_MAX_CHARS" (in DexRun.c) is more like 64 bytes per packet. Testing shows that 128 is too much. 

Because the packet receive nulls out the ';' character (which apparently serves as an EOL marker for data packets?) you can't send a ';' in any data from DDE to DexRun. It causes the program to crash. I can't change it because other routines depend on a null at the end of the recieved data. So I had to add a routine that unescapes data on the DexRun side. You can only send ; as %3B. So DDE needs to escape ; to %3B and because '%' is now an escape character, that needs to be escaped as %25. That's the bad news. 

The good news is that we can now send binary data as well, or at least nulls and FFs work, e.g. %00 puts a null in a file and %FF puts an FF in the file.

The bad news is that because each packet has to be escaped individually (so that you don't have a situation where a ; ends up on the last character of a packet and the escaping gets split into two packets and so the unescaping misses it... first packet ends with %, second starts with 3B) the escaping must be done AFTER the data is broken into parts. So if a part happened to be entirely made of ';' it could end up being 3 times longer, which would make a 64 byte packet into a 192 bytes packet which is too big. 

The good news is that it should be possible to send that escaped "too long" packet as three packets.

The bad news is that Dexter.write_to_robot just got very complex.  

Long story short, you can transfer files as big as you like with Dexter.write_to_robot(data, pathfile); as soon as you update DexRun.c to latest in repo and update  max_content_chars to 64 in DDE... And never send a ';' or a '%' or always send them at the right point. LOL. An updated Dexter.write_to_robot that escapes the data and breaks it into packets without ever straddling an escape will be needed.

Test program sends 250 bytes:
````
var line = "This%00line is 25 byte%25";
var data = "";
for (var i = 0; i < 10; i++) { //10 lines gives us 250 bytes of data. 
	data += line + ("000" + i).substr(-3) + "%3B\n";
	}
out(data);
out(data.length);
new Job({name: "my_job",
         do_list: [Dexter.write_to_robot(data, "/srv/samba/share/test.txt")]})
````

Notice this works only because the data is escaped in such a way that it never happens to hit a packet boundary.