Make JSON ok for SetParam Ctrl parameters
Add '[', ']', and ':' to the list of separators for the SetParam Ctrl parsing so that JSON like syntax is ignored. E.g. instead of having to do 
PID_P 1,2,3 
you can do 
[PID_P: 1, 2, 3] 
and the firmware won't see any difference. This just makes it easy to use JSON commands in Javascript to build up the command. 

See:

https://github.com/HaddingtonDynamics/Dexter/wiki/set-parameter-oplet#ctrl

https://github.com/HaddingtonDynamics/Dexter/issues/33#issuecomment-418529231

https://github.com/HaddingtonDynamics/Dexter/commit/f0d9fa772ba6c3eee979e62a071bca487a084c21