Fix bug where closed file handle was not null.
Using read_from_robot, then using write_to_robot caused DexRun to crash.

It turns out that fclose is setting the re-used file handle variable, wfp, to -1! The standard says that the value for the file pointer is "undefined" after fclose, so we can't depend on it to be valid, 
https://en.cppreference.com/w/c/io/fclose
which we were doing, because we were testing it for null to see if the file needed to be closed. This would be important because write_to_robot doesn't close and re-open the file between each socket full of data. If a write_to_robot was running, and interrupted. E.g. the DDE connection was severed, then when the next write_to_robot was run, the file handle would be overwritten and that would leak memory.

Interesting that there are multiple conflicting answers to the question of how to manage the file handle after the file is closed:
https://stackoverflow.com/questions/8441059/what-happens-to-file-pointer-after-file-is-closed/8441183

So two ways to manage it:
- If we check if wfp is in use and close before opening, as write_to_robot must do, incase a multi-buffer write gets interrupted, then we must check if it is /positive/ instead of non-zero. 
- If we fclose, we should manually set it to zero. Which had been done on write_to_robot, but was not done on read_from_robot and other settings file reads.