Fix RealTimeMonitor ServoRx SendReadPacket bug introduced in last update
An incorrect number of bytes was being read back from the Dynamixel servos by the Real Time Monitor. This bug was introduced by an overzealous attempt to write "good" C code that didn't hardcode lengths and instead used sizeof(ServoRx). In this case, it wasn't needed and set an incorrect value.