Add Monitor Limit setparameters
Add SetParameter 'S' oplet subcommands:

RawEncoderErrorLimits The largest error allowed between the raw encoder reading of position and the expected position for each joint. Part of the status monitor.

RawVelocityLimits The largest velocity allowed between subsequent readings of the raw encoder position. Part of the status monitor.

The values set by these are used in the monitorTorque routine to detect problems. At this point, a debug message is printed, but the next step will be to set DexError which is returned to DDE and then drop to open loop mode.