Torque and Velocity Monitor actions, Faster SetParameter
Set DexError variable for return to DDE of special error code when position error or velocity exceeds a limit. 
#define MONITOR_ERROR_CODE 666

Switch to open loop mode automatically when velocity exceeds limit. 
Add subroutine setOpenLoop().

Change SetParam so that if the second parameter is a number, it will be used as the number of the parameter to set. e.g. "S Speed 20" is the same as "S 0 20". More importantly, "S CartesianSpeedStart 20" can now be written as "S 40 20" and it will be found an executed much faster: The processor will not have to run through an array of 39 other set parameter sub commands doing a strcmp on each one before finding that item. In most cases, this extra time is not an issue, but it is possible that with a job engine running on Dexter, sending many small MOVETOSTRAIGHT "T" oplets, with multiple speed setting commands around each one, we might reach a point where the lookup is a bottleneck. This change was relatively cheap and (hopefully) safe and will help us avoid that issue in the future. There is one small proviso, as noted in the code:
//NEVER ADD A NAME STARTING WITH A NUMERICAL DIGIT!