Add RunDexRun
This is a bash script that checks the file dates on DexRun.c (the source code) and DexRun (the executable) and recompiles the source if it is newer. 

Note that this does NOT require the system date to be set because we use the -p option on the `cp` (file copy) command to retain the file date when copying the source to the project folder for compilation. 

The resulting executable is then `touch`ed with the date of the source file and copied back to the /srv/samba/share folder. This ensures the re-compile only happens once. 

If the compile fails, (e.g. bug in the new source code) the existing executable will still be touched and will be used instead. 

The actions taken are logged into a file `/srv/samba/share/DexRun.log` 

With this system, new firmware updates could potentially be automatically downloaded by DDE, and written to Dexter via write_to_robot, greatly simplifying the update process (e.g. no need for SSH, PuTTY, manual date setting, etc...) The updated source file will have the correct date because DexRun.c (on this branch) now sets the system date from the first data packet it receives from DDE.