Support new Gateware, return #StepAngles, Access Shell
THIS CODE REQUIRES A NEW XILLYDEMO.BIT FILE!
You can get from https://github.com/HaddingtonDynamics/Dexter/commit/c5a1be5854bfa30abe10906dcf2a6a2e9ebd53e8 and follow the directions in the Gateware folder to install it. 

Support new xillydemo.bit file which directly exposes the position of the stepper motors
- Add #Steps and #StepAngles "files" for read_from_robot ('r') oplet. These values represent the position in stepper motor micro steps, of each joints drive motor. StepAngles is converted into integer arcseconds.

Check for mismatches between the Gateware and Firmware
- Move the code for oplet 'r' which reads FPGA addresses into the status reply section so it can return those values over the socket instead of only displaying them on the console. 
- Check the INPUT_OFFSET that DexRun.c supports against the 2nd from the top byte in the first address of the mapped FPGA memory. If that doesn't match: 
- - build a "fake" memory "map" from ram (just an array) and use that so the program can continue running without being able to affect the FPGA at all.
- - Report this mismatch on ALL status responses via variable OverError by setting bit ERROR_INPUT_OFFSET. Bitwise OR OverError with DexError on returned data so that any opcode related error is not lost. E.g. if the Firmware Gateware mismatch is being reported, and the user tries to write a new DexRun.c, but the file system is messed up, pass back both the ERROR_INPUT_OFFSET bit along with whatever error the file system returns. 

Support shelling out to BASH and returns results of Linux commands via 'r' read_from_robot oplet. 
- Add ability to create a child process with the BASH shell and send it a command to execute. This is opened as a file which is then available to be read back from like any other file.  The flag for this is the ` or "backtick" (ASCII 96 0x60) character before the command. 
e.g. "r 0 `ls" will return a directory listing; use "r 1 `" (or "r" with any number >0 and a backtick) repeatedly to get the rest of the data
NOTE: It is critical when using this feature to read all the way to the end until an error 10 is returned. Do not stop sending "r 1 `" oplets just because you get back less than MAX_CONTENT_CHARS data. The reason for this is that the BASH command may require time to complete it's output. It might send back a few bytes at a time and not fill the return buffer completely with every read. When the child process has completed, doing a read will return an "end of stream" error which clearly indicates the process has finished. Until then, the process is running and may continue to generate output. New 'r' shell commands will fail because the file handle is still in use. 
CAUTION: If this function is not used correctly, it can really scramble the system. E.g. you can format the SD card with this. Or you can lose connectivity because the system is waiting on the child process to stop. 

Minor:
- Add in a check for "30" as a synonym for "RunFile" SetParameter command so that it can also be activated via a number instead of name. E.g. S 30 name.make_ins works as well as S RunFile name.make_ins. This was needed because the other parameter names are recognized in a different part of the program, and RunFile is a special case. 
- Cleanup of code in the "#" file section of 'r' for consistency and to remove unneeded debug stuff.