Add shadow_map
Add an array the same size as the mapped memory and update the corresponding address in it every time we write to mapped memory. This allows us to track the last value written since the mapped FPGA memory is write only. 

TODO: Use shadow_map to monitor the status of the END_EFFECTOR_IO setting and NOT send Dynamixel commands if it's not set into a Dynamixel mode.