Don't monitor Dynamixels if they are not in use.
Use the last setting of shadow_map[END_EFFECTOR_IO] to decide if the blue and green wires are setup for Dynamixel io, and if not, don't send position requests in the monitor thread. 

Add some defines for IO_BLUE_ and IO_GREEN_ to make the functions of the bits in END_EFFECTOR_IO clear

Add a #define IO_DYNAMIXEL (IO_BLUE_DYNAMIXEL + IO_GREEN_IN) which is the value set in END_EFFECTOR_IO when the Dyanmixels are being used.