Torque limit joint 6 / 7 servos.
The idea here is to reduce the maximum "goal" torque the Dynamixel servos will use to move to a specific position. By setting this to a value LESS than the MaxTorque (which is the torque at which it throws an error) we should be able to keep pushing without the servo going into an error state.