Correct ANGLE_END_RATIO value
Apparently an error was introduced a long time ago because the robot used to make this (Jame Newtons HD) has nonstandard pulleys on its Joint 4 and 5 axis motors. That was compensated for and then that got copied in here by mistake.  If any Dexter HD's are going to the wrong angle in open loop mode, this is probably why. Closed loop would compensate automatically but would start at a slightly incorrect angle.