Add list of jobs, click to run
This makes use of an updated ability in httpd.js to return a list of files in the dde_apps folder on the robot and to start jobs in the onboard job engine. We make a simple ajax request for the list of files and convert each of them into a link to the runjob url. Click the link to run the job.