No returned errors, MaxTorque for J6/7, servo coms error checking.
Add define for MONITOR_XL320 which will return servo errors to DDE. With this commented out, no errors should be returned to DDE so the LTS version won't stop jobs. Newer versions of DDE will support conditional error handling. 

Remove error checking for Boundaries. Again, to support LTS DDE. 

Add SetParameter for "MaxTorque" which currently only sets the trigger point for load on the servos. Above this load, the "bang/bang" controller will kick in and command the servo to move to where it is (vs the actual goal) in order to limit torque. When the load drops, the actual goal is reinstated, probably causing overtorque again. The system "bang / bangs" between these two extremes but never errors out.

The default torque limit is #define XL320_LOAD_LIMIT 200

Torque units are approx 10 times the percentage of maximum torque. E.g. 100 is 10 percent of the max the servo can do.

This system also checks the error in reported position vs goal and continuously sends the goal position if it isn't met. This will probably help to increase the reliability of the Joints 6 and 7 responding to commands. 

Move coms error checking for servos into the SendReadPacket routine for all servos and don't update ServoData values when bad packets are recieved.