Don't print debug msg if Servo not installed
The existing code would print an error about the returned header being bad every time it tried to communicate with the Dynamixel servos when one or both was not installed. A simple fix is to check if the bytes returned are all the same and if so, just reject the packet, but don't show the error.