Add Websocket Job Engine interface.
Add web socket interface for running job engine programs on Dexter with 2 way communications. 

This is done on a url "/runjob" on the standard www port 80 instead of using a separate port to ease future issues with firewalls. The old use of /runjob?jobname.dde is still supported via standard http or ajax. 

Move job spawn out into it's own function with call back for returned data and status (status meaning process end). This allows it to be used both for ajax and web socket. 

Data sent to stdout by the child, job, process will be returned to the browser via the web socket. Data sent from the browser via the web socket will be forwarded to the stdin of the job process.