define out Dynamixel load limiting
Tag the code that checks for overloading of the Dynamixels and backs off the goal position or checks position error and re-asserts the actual goal for Joint 6 and 7, under the define "MONITOR_XL320" which is currently NOT defined. This has the effect of removing that code from production, but keeping it available for easy reinsertion. 

Don't put it back in until Kent, et all have a chance to test and debug or at least communicate what the problem is with it.