Support Job Engine status and 2 way coms
This implements the changes made by Fry to support a more complete interface between the Job Engine and a users browser. 

It retains the proxy between the raw socket on DexRun and the browser on web socket port 3000, but moves the Job Engine communication to port 3001 to support separate running of a job and direct connection to DexRun at the same time. 

Job Engine job status and data is communicated. show window commands in the job should cause a dialog to appear in the browser. Actions in that dialog will be communicated back to the job just as they would be if the job is running in DDE. 

Note that this needs the new "jobs.html" file and that needs the je_and_browser_code.js file from the dde/core folder. You can copy that file from dde/core to the share/www folder, or use the following command from share/www to create a symlink: 
`ln -s /root/Documents/dde/core/je_and_browser_code.js je_and_browser_code.js`